Move dump_frames to otx/api/utils
