+++ otx/algorithms/common/utils/logger.py
@@ -11,6 +11,7 @@ import os
 import sys
 from typing import Callable
 
+import torch.distributed as dist
 
 # __all__ = ['config_logger', 'get_log_dir', 'get_logger']
 __all__ = ["config_logger", "get_log_dir"]
@@ -115,7 +116,6 @@ class _DummyLogger(logging.Logger):
 
 
 def local_master_only(func: Callable) -> Callable:
-    import torch.distributed as dist
     """A decorator that allows a function to be executed only by the local master process in distributed training setup.
 
     Args:
