+++ otx/cli/tools/demo.py
@@ -27,7 +27,7 @@ from otx.api.entities.inference_parameters import InferenceParameters
 from otx.api.entities.task_environment import TaskEnvironment
 from otx.cli.manager import ConfigManager
 from otx.cli.tools.utils.demo.images_capture import open_images_capture
-from otx.cli.tools.utils.demo.visualization import draw_predictions, dump_frames, put_text_on_rect_bg
+from otx.cli.tools.utils.demo.visualization import draw_predictions, put_text_on_rect_bg
 from otx.cli.utils.importing import get_impl_class
 from otx.cli.utils.io import read_label_schema, read_model
 from otx.cli.utils.parser import (
